requirements.txt
Flask
mysql-connector-python
Flask-WTF
Werkzeug